// app/utils/endpoints/types.ts
import type { Tier } from "../tier";

export interface IApiAccountResponse {
   id: string;
   gameName: string;
   tagLine: string;
   region: string;
}

interface IApiProfileSummoner {
   profileIcon: number;
   level: number;
}

export interface IApiProfileResponse extends IApiAccountResponse {
   summoner: IApiProfileSummoner;
}

interface IApiChallengeSummoner extends IApiProfileSummoner {
   tier: Tier;
   totalPoints: number;
   percentile: number;
   title: string | null;
   displayedChallenges?: number[];
}

interface IApiChallenge {
   challengeId: number;
   percentile?: number;
   tier: Tier;
   points: number;
   achievedTime?: number;
   position?: number;
   playersInLevel?: number;
}

export interface IApiChallengeResponse extends IApiProfileResponse {
   summoner: IApiChallengeSummoner;
   challenges: IApiChallenge[];
}

export interface IApiVerifiedResponse {}